pyproject.toml
[project]
name = "omrchecker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepmerge>=1.1.0",
    "dotmap>=1.3.30",
    "flake8>=6.0.0",
    "freezegun>=1.2.2",
    "jsonschema>=4.17.3",
    "matplotlib>=3.7.1",
    "numpy>=1.25.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.0.2",
    "pillow>=10.0.0",
    "pyside6>=6.5.0",
    "pre-commit>=3.3.3",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
    "rich>=13.4.2",
    "screeninfo>=0.8.1",
    "syrupy>=4.0.4",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
